Restructure packaging for gift card
COAND-497
