+++ giftcard/src/main/java/com/adyen/checkout/giftcard/internal/ui/GiftCardViewProvider.kt
@@ -6,7 +6,7 @@
  * Created by josephj on 30/9/2022.
  */
 
-package com.adyen.checkout.giftcard
+package com.adyen.checkout.giftcard.internal.ui
 
 import android.content.Context
 import android.util.AttributeSet
@@ -14,6 +14,8 @@ import com.adyen.checkout.components.ui.ComponentView
 import com.adyen.checkout.components.ui.ViewProvider
 import com.adyen.checkout.components.ui.view.ButtonComponentViewType
 import com.adyen.checkout.components.ui.view.ComponentViewType
+import com.adyen.checkout.giftcard.R
+import com.adyen.checkout.giftcard.internal.ui.view.GiftCardView
 
 internal object GiftCardViewProvider : ViewProvider {
