+++ giftcard/src/test/java/com/adyen/checkout/giftcard/internal/util/GiftCardBalanceUtilsTest.kt
@@ -1,8 +1,6 @@
-package com.adyen.checkout.giftcard
+package com.adyen.checkout.giftcard.internal.util
 
 import com.adyen.checkout.components.model.payments.Amount
-import com.adyen.checkout.giftcard.util.GiftCardBalanceStatus
-import com.adyen.checkout.giftcard.util.GiftCardBalanceUtils
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
